move service creation into drb invoker
